% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_time.R
\name{edit_time}
\alias{edit_time}
\title{A small function for string->datetime conversion.}
\usage{
edit_time(x)
}
\arguments{
\item{x}{The string from the time field.}
}
\value{
A datetime.
}
\description{
A small function for string->datetime conversion.
}
\examples{
edit_time("2020-01-01T10:00:00Z")

}
