% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station.R
\name{get_station}
\alias{get_station}
\title{Get the air quality and air pollution data from a station}
\usage{
get_station(id)
}
\arguments{
\item{id}{The ID of the station (Run stations() function to get these)}
}
\value{
A tibble with all the measurements available.
}
\description{
Get the air quality and air pollution data from a station
}
\examples{
get_station("125707")

}
