% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Get all the stations in your chosen area.}
\usage{
stations(llon, rlon, ulat, dlat)
}
\arguments{
\item{llon}{Longitude of the left boundary}

\item{rlon}{Longitude of the right boundary}

\item{ulat}{Latitude of the upper boundary}

\item{dlat}{Latitude of the lower boundary}
}
\value{
Returns a data.frame with basic informations about the stations.
}
\description{
Get all the stations in your chosen area.
}
\examples{
stations(llon=25.93, rlon=26.2, ulat=44.64, dlat=44.30)

}
